requirements/base.txt
Django==1.6.4
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2
django-widget-tweaks==1.3
django-grappelli==2.5.3
django-easy-pdf==0.1.0
django-babel==0.3.4
django-extensions==1.3.9
psycopg2==2.5.3
django-debug-toolbar==1.2


